# contracts/openapi/public-service.json
{
  "components": {
    "schemas": {
...
      "EventDisplayResponse": {
        "properties": {
          "course_display": {
            "title": "Course Display",
            "type": "string"
          },
          "display_label": {
            "title": "Display Label",
            "type": "string"
          },
          "family_name": {
            "title": "Family Name",
            "type": "string"
          },
          "ordinal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ordinal"
          }
        },
        "required": [
          "course_display",
          "family_name",
          "ordinal",
          "display_label"
        ],
        "title": "EventDisplayResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ManualEventMutationResponse": {
        "properties": {
          "applied": {
            "title": "Applied",
            "type": "boolean"
          },
          "change_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Id"
          },
          "entity_uid": {
            "title": "Entity Uid",
            "type": "string"
          },
          "event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManualEventResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "idempotent": {
            "title": "Idempotent",
            "type": "boolean"
          },
          "lifecycle": {
            "enum": [
              "active",
              "removed"
            ],
            "title": "Lifecycle",
            "type": "string"
          }
        },
        "required": [
          "applied",
          "idempotent",
          "change_id",
          "entity_uid",
          "lifecycle"
        ],
        "title": "ManualEventMutationResponse",
        "type": "object"
      },
      "ManualEventResponse": {
        "properties": {
          "course_dept": {
            "title": "Course Dept",
            "type": "string"
          },
          "course_display": {
            "title": "Course Display",
            "type": "string"
          },
          "course_number": {
            "title": "Course Number",
            "type": "integer"
          },
          "course_quarter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Quarter"
          },
          "course_suffix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Suffix"
          },
          "course_year2": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Year2"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "due_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Time"
          },
          "entity_uid": {
            "title": "Entity Uid",
            "type": "string"
          },
          "event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserFacingEventResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "event_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Name"
          },
          "family_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Id"
          },
          "family_name": {
            "title": "Family Name",
            "type": "string"
          },
          "lifecycle": {
            "enum": [
              "active",
              "removed"
            ],
            "title": "Lifecycle",
            "type": "string"
          },
          "manual_support": {
            "title": "Manual Support",
            "type": "boolean"
          },
          "ordinal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ordinal"
          },
          "raw_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Type"
          },
          "time_precision": {
            "anyOf": [
              {
                "enum": [
                  "date_only",
                  "datetime"
                ],
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "Time Precision"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "course_display",
          "course_dept",
          "course_number",
          "entity_uid",
          "lifecycle",
          "manual_support",
          "family_name",
          "time_precision",
          "created_at",
          "updated_at"
        ],
        "title": "ManualEventResponse",
        "type": "object"
      },
      "ManualEventWriteRequest": {
        "additionalProperties": false,
        "properties": {
          "due_date": {
            "format": "date",
            "title": "Due Date",
            "type": "string"
          },
          "due_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Time"
          },
          "event_name": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Event Name",
            "type": "string"
          },
          "family_id": {
            "minimum": 1.0,
            "title": "Family Id",
            "type": "integer"
          },
          "ordinal": {
            "anyOf": [
              {
                "maximum": 999.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ordinal"
          },
          "raw_type": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Type"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "time_precision": {
            "default": "datetime",
            "enum": [
              "date_only",
              "datetime"
            ],
            "title": "Time Precision",
            "type": "string"
          }
        },
        "required": [
          "family_id",
          "event_name",
          "due_date"
        ],
        "title": "ManualEventWriteRequest",
        "type": "object"
      },
...
      "UserFacingEventResponse": {
        "properties": {
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "due_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Time"
          },
          "event_display": {
            "$ref": "#/components/schemas/EventDisplayResponse"
          },
          "time_precision": {
            "title": "Time Precision",
            "type": "string"
          },
          "uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uid"
          }
        },
        "required": [
          "event_display",
          "time_precision"
        ],
        "title": "UserFacingEventResponse",
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "calendar_delay_seconds": {
            "title": "Calendar Delay Seconds",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "notify_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notify Email"
          },
          "timezone_name": {
            "title": "Timezone Name",
            "type": "string"
          },
          "timezone_source": {
            "title": "Timezone Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "notify_email",
          "timezone_name",
          "timezone_source",
          "calendar_delay_seconds",
          "created_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "calendar_delay_seconds": {
            "anyOf": [
              {
                "maximum": 3600.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Delay Seconds"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "notify_email": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notify Email"
          },
          "timezone_name": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone Name"
          },
          "timezone_source": {
            "anyOf": [
              {
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone Source"
          }
        },
        "title": "UserUpdateRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CalendarDIFF API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/manual/events": {
      "get": {
        "operationId": "get_manual_events_manual_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_removed",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Removed",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ManualEventResponse"
                  },
                  "title": "Response Get Manual Events Manual Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Manual Events",
        "tags": [
          "manual"
        ]
      },
      "post": {
        "operationId": "post_manual_event_manual_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualEventWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualEventMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Manual Event",
        "tags": [
          "manual"
        ]
      }
    },
    "/manual/events/{entity_uid}": {
      "delete": {
        "operationId": "remove_manual_event_manual_events__entity_uid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entity_uid",
            "required": true,
            "schema": {
              "title": "Entity Uid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualEventMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Manual Event",
        "tags": [
          "manual"
        ]
      },
      "patch": {
        "operationId": "patch_manual_event_manual_events__entity_uid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entity_uid",
            "required": true,
            "schema": {
              "title": "Entity Uid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualEventWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualEventMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Manual Event",
        "tags": [
          "manual"
        ]
      }
    },
...
    "/settings/profile": {
      "get": {
        "operationId": "get_profile_settings_profile_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "patch_profile_settings_profile_patch",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Profile",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}